update job submit

+++ submit_job.sh
@@ -4,8 +4,10 @@ set -euo pipefail
 
 job_name=$1
 config_file=$2
-sync_code=${3:-"true"}
-
+cleanup=${3:-"false"}
+dry_run=${4:-"false"}
+sync_code=${5:-"true"}
+config_name=$(basename "${config_file%.*}")
 echo "Job: $job_name $config_file"
 
 if [ -z "$job_name" ] || [ -z "$config_file" ]; then
@@ -16,6 +18,14 @@ if [ "$sync_code" == "true" ]; then
     echo "syncing code to $job_name"
     rcall-brix sync "$job_name"
 fi
+if [ "$cleanup" == "true" ]; then
+    echo "cleaning up old ${config_name} on $job_name"
+    rcall-brix ssh "$job_name" "bash -lc \"python /root/code/verl/ray_job.py cleanup ${config_name}\""
+fi
 
-config_name=$(basename "${config_file%.*}")
+if [ "$dry_run" == "true" ]; then
+    echo "Dry run"
+    echo rcall-brix ssh  "$job_name" "bash -l /root/code/verl/quick_run.sh $config_file" | tee "${job_name}_${config_name}.log"
+    exit 0
+fi
 rcall-brix ssh  "$job_name" "bash -l /root/code/verl/quick_run.sh $config_file" | tee "${job_name}_${config_name}.log"
